requirements.txt
swibots
requests
pillow